fix: update doctor imports after apple consolidation
